test(frameworks): move keras_tf2 tests to test_frameworks
