+++ scripts/ci/config.yml
@@ -102,8 +102,7 @@ keras_tf1:
     - "h5py==2.10.0"
 
 keras_tf2:
-  <<: *tmpl
-  override_name_or_path: test_keras_impl.py
+  <<: *ntmpl
   dependencies:
     - "protobuf<3.20,>=3.9.2"
     - "tensorflow==2.7.3"
